Remove widget_info.x
